fix: rename HistoryList
